:bug: Fix #1 modify file permissions to give system user access
Build the WordPress image with a Dockerfile in order to run
commands to change the owner of the container files to the system
user so that both the user and WordPress can have write access.

Also create a new `start.sh` script to handling the usual `up`
command to set the correct user and group IDs for the Dockerfile
to use.

Set up persistent volumes for the WordPress application and the
plugin and theme directories for simpler development access.